Move aggregates model to its own folder
