+++ src/EventSourced.Diagnostics.Web/Pages/AggregateTypes/AggregateTypesListViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using DotVVM.Framework.ViewModel;
-using EventSourced.Diagnostics.Web.Model;
+using EventSourced.Diagnostics.Web.Model.Aggregates;
 using EventSourced.Diagnostics.Web.Services;
 
 namespace EventSourced.Diagnostics.Web.Pages.AggregateTypes
